requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
APScheduler==3.10.4
attrs==23.2.0
Babel==2.14.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
greenlet==3.0.3
gspread==6.1.0
idna==3.6
magic-filter==1.0.12
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
oauthlib==3.2.2
pillow==10.2.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.5.3
pydantic-settings==2.2.1
pydantic_core==2.14.6
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
segno==1.6.1
SQLAlchemy==2.0.27
StrEnum==0.4.15
typing_extensions==4.9.0
urllib3==2.2.1
XlsxWriter==3.2.0
yarl==1.9.4